Imported Some Assets
So a long time ago I did this pause menu thing that had a whole way to look at settings and an audio adjustment tap and it worked and stuff. I'm not sure how useful it'd actually be but if you want me to finish it let me know. It's supposed to have a prefab but I deleted it because it was unfinished and erroring and I can't publish a change that has errors so again if you think it's useful to the point where I should finish it let me know.